fix: Relocate 'Starting Snippets' feature to viewer
This commit corrects the implementation of the 'Starting Snippets' feature based on user clarification.

The feature was incorrectly implemented in the `SnippetEditor` component. It has now been moved to the `SnippetViewer` component, which is the correct place for composing the final output text.

Changes:
- The dropdown menu for selecting a starting snippet has been removed from the editor modal.
- The dropdown menu has been added to the viewer page, where placeholders are filled.
- Selecting a starting snippet now prepends its content to the live preview and the final copied text.
