+++ kolder-app/src/components/SnippetEditor.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef } from 'react';
+import { useState, useEffect } from 'react';
 import {
   Modal,
   ModalOverlay,
@@ -11,34 +11,16 @@ import {
   FormLabel,
   Input,
   Button,
-  Select,
   VStack,
 } from '@chakra-ui/react';
 import ReactQuill from 'react-quill';
 import 'react-quill/dist/quill.snow.css';
-import './quill.css'; // Import custom Quill styles
-import axios from 'axios';
-
-const api = axios.create({
-    baseURL: '/api',
-});
+import './quill.css';
 
 const SnippetEditor = ({ isOpen, onClose, onSave, snippet, settings }) => {
   const [name, setName] = useState('');
   const [content, setContent] = useState('');
-  const [startingSnippets, setStartingSnippets] = useState([]);
-  const quillRef = useRef(null); // Ref to access Quill editor instance
 
-  // Fetch starting snippets when modal opens
-  useEffect(() => {
-    if (isOpen) {
-        api.get('/starting-snippets')
-           .then(response => setStartingSnippets(response.data))
-           .catch(error => console.error('Error fetching starting snippets', error));
-    }
-  }, [isOpen]);
-
-  // Populate editor when snippet is being edited
   useEffect(() => {
     if (isOpen) {
         if (snippet) {
@@ -56,21 +38,6 @@ const SnippetEditor = ({ isOpen, onClose, onSave, snippet, settings }) => {
     onClose();
   };
 
-  const handleStartingSnippetChange = (e) => {
-    const snippetId = e.target.value;
-    const selected = startingSnippets.find(ss => ss._id === snippetId);
-    if (selected) {
-        // Prepend the content
-        const newContent = selected.content + '<p><br></p>' + content;
-        setContent(newContent);
-
-        // Focus the editor
-        if (quillRef.current) {
-            quillRef.current.getEditor().focus();
-        }
-    }
-  };
-
   return (
     <Modal isOpen={isOpen} onClose={onClose} size="xl">
       <ModalOverlay />
@@ -78,23 +45,15 @@ const SnippetEditor = ({ isOpen, onClose, onSave, snippet, settings }) => {
         <ModalHeader>{snippet ? 'Edit Snippet' : 'New Snippet'}</ModalHeader>
         <ModalCloseButton />
         <ModalBody>
-            <VStack spacing={4} align="stretch">
-                <FormControl>
-                    <FormLabel>Name</FormLabel>
-                    <Input value={name} onChange={(e) => setName(e.target.value)} />
-                </FormControl>
-                <FormControl>
-                    <FormLabel>Insert Starting Snippet</FormLabel>
-                    <Select placeholder="Select a starting snippet..." onChange={handleStartingSnippetChange}>
-                        {startingSnippets.map(ss => (
-                            <option key={ss._id} value={ss._id}>{ss.name}</option>
-                        ))}
-                    </Select>
-                </FormControl>
-                <FormControl>
-                    <FormLabel>Content</FormLabel>
-                    <ReactQuill ref={quillRef} theme="snow" value={content} onChange={setContent} />
-                </FormControl>
+          <VStack spacing={4} align="stretch">
+            <FormControl>
+                <FormLabel>Name</FormLabel>
+                <Input value={name} onChange={(e) => setName(e.target.value)} />
+            </FormControl>
+            <FormControl>
+                <FormLabel>Content</FormLabel>
+                <ReactQuill theme="snow" value={content} onChange={setContent} />
+            </FormControl>
           </VStack>
         </ModalBody>
         <ModalFooter>
