+++ kolder-app/src/components/SnippetViewer.jsx
@@ -9,6 +9,7 @@ import {
   Input,
   Text,
   VStack,
+  Select,
 } from '@chakra-ui/react';
 import { ArrowBackIcon } from '@chakra-ui/icons';
 import axios from 'axios';
@@ -21,24 +22,17 @@ const SnippetViewer = ({ snippet, onBack, settings }) => {
   const [placeholders, setPlaceholders] = useState({});
   const [output, setOutput] = useState('');
   const [hasCopied, setHasCopied] = useState(false);
+  const [startingSnippets, setStartingSnippets] = useState([]);
+  const [prefix, setPrefix] = useState('');
 
-  const handleCopy = () => {
-    // 1. Create a temporary element to parse the HTML
-    const tempDiv = document.createElement('div');
-    tempDiv.innerHTML = output;
-    // 2. Get the plain text content
-    const plainText = tempDiv.textContent || tempDiv.innerText || '';
+  // Fetch starting snippets
+  useEffect(() => {
+    api.get('/starting-snippets')
+       .then(response => setStartingSnippets(response.data))
+       .catch(error => console.error('Error fetching starting snippets', error));
+  }, []);
 
-    // 3. Use the Clipboard API to copy the plain text
-    navigator.clipboard.writeText(plainText).then(() => {
-        setHasCopied(true);
-        setTimeout(() => setHasCopied(false), 2000); // Reset after 2 seconds
-    });
-
-    // 4. Track the usage via API
-    api.post(`/snippets/${snippet._id}/track`).catch(err => console.error("Failed to track copy", err));
-  }
-
+  // Set placeholders when snippet changes
   useEffect(() => {
     if (!snippet.content) return;
     const placeholderRegex = /{{(.*?)}}/g;
@@ -48,18 +42,43 @@ const SnippetViewer = ({ snippet, onBack, settings }) => {
         initialPlaceholders[match[1]] = '';
     });
     setPlaceholders(initialPlaceholders);
+    setPrefix(''); // Reset prefix when snippet changes
   }, [snippet]);
 
+  // Update the final output when prefix or placeholders change
   useEffect(() => {
-    let newOutput = snippet.content || '';
+    let snippetWithPlaceholders = snippet.content || '';
     for (const key in placeholders) {
-      newOutput = newOutput.replace(new RegExp(`{{${key}}}`, 'g'), placeholders[key]);
+      snippetWithPlaceholders = snippetWithPlaceholders.replace(new RegExp(`{{${key}}}`, 'g'), placeholders[key]);
     }
-    setOutput(newOutput);
-  }, [placeholders, snippet]);
+    setOutput(prefix + snippetWithPlaceholders);
+  }, [prefix, placeholders, snippet]);
+
+  const handleCopy = () => {
+    const tempDiv = document.createElement('div');
+    tempDiv.innerHTML = output;
+    const plainText = tempDiv.textContent || tempDiv.innerText || '';
+    navigator.clipboard.writeText(plainText).then(() => {
+        setHasCopied(true);
+        setTimeout(() => setHasCopied(false), 2000);
+    });
+    api.post(`/snippets/${snippet._id}/track`).catch(err => console.error("Failed to track copy", err));
+  }
 
   const handlePlaceholderChange = (key, value) => {
     setPlaceholders(prev => ({ ...prev, [key]: value }));
+  };
+
+  const handleStartingSnippetChange = (e) => {
+    const snippetId = e.target.value;
+    if (!snippetId) {
+        setPrefix('');
+        return;
+    }
+    const selected = startingSnippets.find(ss => ss._id === snippetId);
+    if (selected) {
+        setPrefix(selected.content + '<p><br></p>');
+    }
   };
 
   return (
@@ -73,19 +92,29 @@ const SnippetViewer = ({ snippet, onBack, settings }) => {
 
       <VStack spacing="4" align="stretch">
         <Box bg={settings?.theme.contentBackgroundColor} p={4} borderRadius="md">
-          <Heading size="sm" mb="2">Fill Placeholders</Heading>
-          {Object.keys(placeholders).length > 0 ? (
-            Object.keys(placeholders).map(key => (
-              <FormControl key={key} mt="2">
-                <FormLabel>{key}</FormLabel>
-                <Input
-                  value={placeholders[key]}
-                  onChange={(e) => handlePlaceholderChange(key, e.target.value)}
-                />
-              </FormControl>
-            ))
-          ) : (
-            <Text>No placeholders in this snippet.</Text>
+          <Heading size="sm" mb="2">Compose Snippet</Heading>
+          <FormControl>
+            <FormLabel>Add Starting Snippet</FormLabel>
+            <Select placeholder="None" onChange={handleStartingSnippetChange}>
+                {startingSnippets.map(ss => (
+                    <option key={ss._id} value={ss._id}>{ss.name}</option>
+                ))}
+            </Select>
+          </FormControl>
+
+          {Object.keys(placeholders).length > 0 && (
+            <Box mt={4}>
+                <Heading size="sm" mb="2">Fill Placeholders</Heading>
+                {Object.keys(placeholders).map(key => (
+                <FormControl key={key} mt="2">
+                    <FormLabel>{key}</FormLabel>
+                    <Input
+                    value={placeholders[key]}
+                    onChange={(e) => handlePlaceholderChange(key, e.target.value)}
+                    />
+                </FormControl>
+                ))}
+            </Box>
           )}
         </Box>
 
@@ -95,6 +124,7 @@ const SnippetViewer = ({ snippet, onBack, settings }) => {
             p="4"
             borderWidth="1px"
             borderRadius="md"
+            minHeight="150px"
             bg={settings?.theme.contentBackgroundColor}
             borderColor={settings?.theme.accentColor}
             dangerouslySetInnerHTML={{ __html: output }}
